the hasLineTerminator logic was faulty; the code had been stop-gapped for template strings, but then the parsing error had been introduced as the stop-gap would incorrectly regard this:
```
return
  `template string`
```

as a single return statement. Lateron a multi-line string stop-gap was added, which was done in the same vain and thus suffered from the same error.

Now test cases have been added to prevent this from happening again and it looks like the code passes, now that we have a more proper ASI complaint hasLineTerminator logic in the parser/scanner interface code at `nextToken()`.